+++ subcommand/controller/command_ent_test.go
@@ -40,9 +40,21 @@ func TestUpsertAnonymousTokenPolicyEnt(t *testing.T) {
 }
 
 func TestUpsertAPIGatewayTokenPolicyAndRole(t *testing.T) {
-	testUpsertAPIGatewayPolicyAndRole(t, map[string]apiGatewayTokenTest{
+	testUpsertAPIGatewayPolicyAndRole(t, map[string]gatewayTokenTest{
 		"test creation": {
 			partitionsEnabled: true,
+		},
+	})
+}
+
+func TestUpsertMeshGatewayTokenPolicyAndRole(t *testing.T) {
+	testUpsertMeshGatewayPolicyAndRole(t, map[string]gatewayTokenTest{
+		"test creation": {
+			partitionsEnabled: true,
+		},
+		"test creation with non default partition": {
+			partitionsEnabled:      true,
+			useNonDefaultPartition: true,
 		},
 	})
 }